refactor(graduate): StudentAddModal 구조 개선 및 상수, 타입, 유틸리티 함수 디렉토리 변경
