--- apps/graduate/src/widgets/StudentAddModal/models/index.ts
@@ -1,3 +1,0 @@
-export { default as StudentAddSingle } from './single/StudentAddSingle';
-export { default as StudentAddMultiple } from './multiple/StudentAddMultiple';
-
