+++ apps/graduate/src/widgets/StudentAddModal/ui/StudentAddMultiple.tsx
@@ -1,8 +1,8 @@
 import { Button, Typography, Upload } from 'antd';
 
 import PreviewTable from './PreviewTable';
-import { useStudentAddMultiple } from './useStudentAddMultiple';
-import type { MultipleUploadRow } from '../../types';
+import { useStudentAddMultiple } from '../model/useStudentAddMultiple';
+import type { MultipleUploadRow } from '../types/studentAddModal';
 
 type Props = {
   open?: boolean;
